refactor: 모든 컴포넌트 리팩토링
