+++ Frontend/src/pages/FeedView/style.css
@@ -1,0 +1,195 @@
+.feedview {
+  background-color: white;
+  width: 700px;
+  margin: 10px auto;
+  display: inline-block;
+  overflow: hidden;
+}
+
+.post__text {
+  position: relative;
+  margin-top: -2px;
+  font-weight: normal;
+  height: 10%;
+  padding: 3px 10px;
+  text-align: right;
+}
+.feedview_header {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  padding: 5px 10px;
+  background-color: white;
+}
+.feed_click :hover {
+  cursor: pointer;
+}
+
+.feed_avatar {
+  border-radius: 10px;
+  margin-top: 5px;
+  margin-right: 10px;
+  width: 54px;
+  height: 54px;
+  object-fit: contain;
+}
+
+.feed_avatar img {
+  width: 100% !important;
+  height: 100% !important;
+  border-radius: 10px;
+}
+
+.post_header h3 {
+  position: relative;
+  margin-top: -10px;
+}
+
+.timeline {
+  margin: 0 auto;
+  max-width: 500px;
+  padding: 20px;
+}
+
+.clickable {
+  cursor: pointer;
+  margin-left: 0;
+}
+
+.clickable img {
+  max-width: 24px;
+  max-height: 24px;
+}
+
+.post_headerinfo {
+  position: relative;
+  margin-top: -10px;
+  margin-left: 20px;
+  margin-right: 20px;
+}
+
+.icon_anime1 :hover {
+  cursor: pointer;
+}
+
+.icon_anime2 {
+  cursor: pointer;
+  animation: zoomin_out 0.4s;
+  animation-fill-mode: forwards;
+}
+
+@keyframes zoomin_out {
+  0% {
+    transform: scale(1.2, 1.2);
+  }
+  50% {
+    transform: scale(1.1, 1.1);
+  }
+  100% {
+    transform: scale(1, 1);
+  }
+}
+
+@keyframes zoomout_in {
+  100% {
+    transform: scale(1.2, 1.2);
+  }
+  50% {
+    transform: scale(1.1, 1.1);
+  }
+  0% {
+    transform: scale(1, 1);
+  }
+}
+
+div::-webkit-scrollbar {
+  display: none;
+}
+
+textarea::-webkit-scrollbar {
+  display: none;
+}
+
+.comment {
+  padding: 10px 15px;
+  width: 100%;
+  border-color: #ccc #ddd #ddd #ccc;
+}
+
+.comment textarea {
+  width: 100%;
+  margin-top: 5px;
+  border-color: #ccc #ddd #ddd #ccc;
+  border-radius: 3px;
+  box-shadow: inset 1px 1px 4px rgb(0 0 0 /6%);
+  padding: 5px;
+  font-size: 15px;
+}
+
+.comment textarea::placeholder {
+  font-size: 15px;
+}
+
+.comment_list {
+  background-color: white;
+  border-top: 1px solid lightgray;
+  width: 700px;
+  margin: 5px auto;
+  display: block;
+  overflow: hidden;
+}
+
+.commenter {
+  background-color: white;
+  font-size: 15px;
+  width: 700px;
+  margin: 0px auto;
+  display: flex;
+  overflow: hidden;
+}
+.commenter img {
+  width: 50px;
+  height: 50px;
+  border-radius: 5px;
+  margin: 10px 5px;
+  overflow: hidden;
+}
+
+.commenter h5 {
+  position: relative;
+  margin-top: 8px;
+  margin-left: 3px;
+}
+
+._a9yh {
+  align-items: stretch;
+  border: 0;
+  border-bottom: 1px solid rgb(178, 178, 178);
+  box-sizing: border-box;
+  display: inline-block;
+  flex-direction: column;
+  flex-shrink: 0;
+  font: inherit;
+  font-size: 100%;
+  height: 0;
+  margin: 0 8px;
+  margin-top: 8px;
+  padding: 0;
+  position: relative;
+  vertical-align: middle;
+  width: 24px;
+}
+
+.comment_context {
+  max-width: 640px;
+  font-size: 15px;
+  padding: 3px;
+}
+
+.comment_context p {
+  word-wrap: break-word;
+}
+
+.comment_button img {
+  margin: 10px 0px 0 5px;
+}
