+++ Frontend/src/pages/index.js
@@ -1,0 +1,4 @@
+export * from "./FeedView";
+export * from "./Login";
+export * from "./Profile";
+export * from "./Thread";
